Move controllers to paperscape folder
