+++ src/scenes/papers/components/headers/Search.js
@@ -1,8 +1,8 @@
 /* encoding: utf-8 */
 
 import React, { Component } from "react";
-import PaperSearchCtl from "../../controllers/PaperSearch";
-import PaperSearchPositionCtl from "../../controllers/PaperSearchPosition";
+import PaperSearchCtl from "../../controllers/paperscape/PaperSearch";
+import PaperSearchPositionCtl from "../../controllers/paperscape/PaperSearchPosition";
 import { Button, Dropdown, Icon, Image, Input, Menu, Segment } from "semantic-ui-react";
