+++ src/scenes/papers/controllers/paperscape/PaperSearch.js
@@ -1,6 +1,6 @@
 /* encoding: utf-8 */
 
-import config from "../../../config"
+import config from "../../../../config"
 
 
 const paperSearchRespPrefix = "(";
